fixed new issues found in clang-tidy check

+++ test/form/generate_vector.cpp
@@ -19,7 +19,7 @@ public:
   {
     std::vector<int> randoms(m_n_time_ticks);
     for (auto& random : randoms)
-      random = m_dist(m_gen);
+      random = static_cast<int>(m_dist(m_gen));
     return randoms;
   }
 
@@ -31,8 +31,8 @@ private:
 
 PHLEX_REGISTER_PROVIDERS(graph, config)
 {
-  int const seed =
-    config.get<int>("seed", std::chrono::system_clock::now().time_since_epoch().count());
+  int const seed = config.get<int>(
+    "seed", static_cast<int>(std::chrono::system_clock::now().time_since_epoch().count()));
   int const n_time_ticks = config.get<int>("n_time_ticks");
   float const mean = config.get<float>("mean", 0);
   float const stddev = config.get<float>("stddev", 1);
